Adjust post refactor
